Separate Markdown processing and Vector search functionality
